+++ R/count.R
@@ -32,13 +32,16 @@ count..data.frame <- function(.df, ...) {
 
   .df <- as_tidytable(.df)
 
-  by <- enquos(...)
+  .by <- enquos(...)
 
-  summarize.(.df, N = .N, by = c(!!!by))
+  summarize.(.df, N = .N, .by = c(!!!.by))
 
 }
 
 #' @export
 #' @rdname count.
-dt_count <- count.
+dt_count <- function(.df, ...) {
+  deprecate_soft("0.5.2", "tidytable::dt_count()", "count.()")
 
+  count.(.df, ...)
+}