+++ R/group_split.R
@@ -31,6 +31,8 @@ group_split..data.frame <- function(.df, ..., .keep = TRUE) {
 
   .df <- as_tidytable(.df)
 
+  vec_assert(.keep, logical(), 1)
+
   dots <- enquos(...)
 
   if (length(dots) == 0) {
@@ -46,6 +48,10 @@ group_split..data.frame <- function(.df, ..., .keep = TRUE) {
 
 #' @export
 #' @rdname group_split.
-dt_group_split <- group_split.
+dt_group_split <- function(.df, ..., .keep = TRUE) {
+  deprecate_soft("0.5.2", "tidytable::dt_group_split()", "group_split.()")
+
+  group_split.(.df, ..., .keep = .keep)
+}