+++ R/mutate.R
@@ -5,7 +5,8 @@
 #'
 #' @param .df A data.frame or data.table
 #' @param ... Columns to add/modify
-#' @param by Columns to group by
+#' @param .by Columns to group by
+#' @param by This argument has been renamed to .by and is deprecated
 #'
 #' @md
 #' @export
@@ -23,24 +24,23 @@
 #' test_df %>%
 #'   mutate.(double_a = a * 2,
 #'           avg_a = mean(a),
-#'           by = c)
-mutate. <- function(.df, ..., by = NULL) {
+#'           .by = c)
+mutate. <- function(.df, ..., .by = NULL, by = NULL) {
   UseMethod("mutate.")
 }
 
 #' @export
-mutate..data.frame <- function(.df, ..., by = NULL) {
+mutate..data.frame <- function(.df, ..., .by = NULL, by = NULL) {
 
   .df <- as_tidytable(.df)
   .df <- shallow(.df)
 
   dots <- enquos(...)
-  by <- enquo(by)
+  .by <- check_dot_by(enquo(.by), enquo(by), "mutate.")
 
-  if (quo_is_null(by)) {
+  if (quo_is_null(.by)) {
     # Faster version if there is no "by" provided
     all_names <- names(dots)
-    data_size <- nrow(.df)
 
     for (i in seq_along(dots)) {
 
@@ -49,37 +49,31 @@ mutate..data.frame <- function(.df, ..., by = NULL) {
 
       # Prevent modify-by-reference if the column already exists in the data.table
       # Fixes cases when user supplies a single value ex. 1, -1, "a"
-      # !is.null(val) allows for columns to be deleted using mutate.(.df, col = NULL)
+      # !quo_is_null(val) allows for columns to be deleted using mutate.(.df, col = NULL)
       if (.col_name %in% names(.df) && !quo_is_null(val)) {
 
         eval_quo(
-          .df[, !!.col_name := eval_quo(
-            {.N = .env$.N; .SD = .env$.SD; .I = .env$.I; .GRP = .env$.GRP;
-            vec_recycle(!!val, data_size)}, .SD)]
+          .df[, !!.col_name := vec_recycle(!!val, .N)]
         )
 
       } else {
 
         eval_quo(
-          .df[, !!.col_name := eval_quo(
-            {.SD = .env$.SD; .N = .env$.N; .I = .env$.I; .GRP = .env$.GRP; !!val},
-            .SD)]
-        , .df)
+          .df[, !!.col_name := !!val]
+        )
       }
     }
   } else {
     # Faster with "by", since the "by" call isn't looped multiple times for each column added
-    by <- select_vec_by(.df, !!by)
+    .by <- select_vec_chr(.df, !!.by)
 
-    dot_names <- names(dots)
-    dots <- unname(dots)
+    needs_copy <- any(vec_in(names(dots), names(.df)))
+
+    if (needs_copy) .df <- copy(.df)
 
     eval_quo(
-      .df[ , !!dot_names := eval_quo(
-        {.N = .env$.N; .SD = .env$.SD; .I = .env$.I; .GRP = .env$.GRP; list(!!!dots)},
-        .SD),
-        by = !!by],
-      .df)
+      .df[, ':='(!!!dots), by = .by]
+    )
 
   }
   .df[]
@@ -87,4 +81,9 @@ mutate..data.frame <- function(.df, ..., by = NULL) {
 
 #' @export
 #' @rdname mutate.
-dt_mutate <- mutate.
+dt_mutate <- function(.df, ..., .by = NULL, by = NULL) {
+  deprecate_soft("0.5.2", "tidytable::dt_mutate()", "mutate.()")
+
+  .by <- check_dot_by(enquo(.by), enquo(by))
+  mutate.(.df, ..., .by = {{ .by }})
+}