+++ R/mutate_across.R
@@ -7,7 +7,8 @@
 #' @param .cols vector `c()` of unquoted column names. `tidyselect` compatible.
 #' @param .fns Functions to pass. Can pass a list of functions.
 #' @param ... Other arguments for the passed function
-#' @param by Columns to group by
+#' @param .by Columns to group by
+#' @param by This argument has been renamed to .by and is deprecated
 #'
 #' @export
 #' @md
@@ -30,26 +31,27 @@
 #' test_df %>%
 #'   mutate_across.(c(x, y), list(new = ~ .x * 2,
 #'                                another = ~ .x + 7))
-mutate_across. <- function(.df, .cols = everything(), .fns, ..., by = NULL) {
+mutate_across. <- function(.df, .cols = everything(), .fns, ..., .by = NULL, by = NULL) {
   UseMethod("mutate_across.")
 }
 
 #' @export
-mutate_across..data.frame <- function(.df, .cols = everything(), .fns, ..., by = NULL) {
+mutate_across..data.frame <- function(.df, .cols = everything(), .fns, ..., .by = NULL, by = NULL) {
 
   .df <- as_tidytable(.df)
 
   .cols <- select_vec_chr(.df, {{ .cols }})
 
-  by <- select_vec_by(.df, {{ by }})
+  .by <- check_dot_by(enquo(.by), enquo(by), "mutate_across.")
+  .by <- select_vec_chr(.df, !!.by)
 
   .df <- shallow(.df)
 
   if (!is.list(.fns)) {
     if (length(.cols) > 0) {
       eval_quo(
-        .df[, (.cols) := eval_quo(map.(.SD, .fns, ...), .SD), .SDcols = .cols, by = !!by],
-        .df)
+        .df[, (.cols) := map.(.SD, .fns, ...), .SDcols = !!.cols, by = !!.by],
+      )
     } else {
       .df
     }
@@ -63,8 +65,8 @@ mutate_across..data.frame <- function(.df, .cols = everything(), .fns, ..., by =
       new_cols <-  paste0(.cols, "_", new_names[[i]])
 
       eval_quo(
-        .df[, (new_cols) := eval_quo(map.(.SD, .fns[[i]]), .SD), .SDcols = .cols, by = !!by],
-        .df)
+        .df[, (new_cols) := map.(.SD, .fns[[i]]), .SDcols = !!.cols, by = !!.by],
+      )
     }
   }
   .df[]
@@ -72,4 +74,9 @@ mutate_across..data.frame <- function(.df, .cols = everything(), .fns, ..., by =
 
 #' @export
 #' @rdname mutate_across.
-dt_mutate_across <- mutate_across.
+dt_mutate_across <- function(.df, .cols = everything(), .fns, ..., .by = NULL, by = NULL) {
+  deprecate_soft("0.5.2", "tidytable::dt_mutate_across()", "mutate_across.()")
+
+  .by <- check_dot_by(enquo(.by), enquo(by))
+  mutate_across.(.df, .cols = {{ .cols }}, .fns, ..., .by = {{ .by }})
+}