+++ R/n.R
@@ -15,7 +15,7 @@
 #'
 #' test_df %>%
 #'   summarize.(count = n.(),
-#'              by = z)
+#'              .by = z)
 #'
 #' test_df %>%
 #'   mutate.(count = n.())
@@ -25,4 +25,8 @@ n. <- function() {
 
 #' @export
 #' @rdname n.
-dt_n <- n.
+dt_n <- function() {
+  deprecate_soft("0.5.2", "tidytable::dt_n()", "n.()")
+
+  n.()
+}