+++ R/utils-general.R
@@ -4,7 +4,20 @@ eval_expr <- function(express) {
 }
 
 eval_quo <- function(express, data = NULL, env = caller_env()) {
-  eval_tidy(enquo(express), data = data, env = env)
+  eval_tidy(quo_squash(enquo(express)), data = data, env = env)
+}
+
+check_dot_by <- function(.by, by, fn) {
+  if (!quo_is_null(by)) {
+
+    if (!missing(fn))
+      deprecate_warn("0.5.2", str_c("tidytable::", fn, "(by = )"), str_c(fn, "(.by = )"))
+
+    by
+
+  } else {
+    .by
+  }
 }
 
 # Creates a shallow copy to prevent modify-by-reference