+++ tests/testthat/test-mutate_variants.R
@@ -1,5 +1,15 @@
 setup(options(lifecycle_verbosity = "quiet"))
 
+test_that("dt_mutate_across(): .cols works with is.numeric & is deprecated", {
+  df <- data.table(x_start = c(1,1,1), end_x = c(2,2,2), z = c("a", "a", "b"))
+  df <- df %>%
+    dt_mutate_across(where(is.numeric), function(.x) .x + 1)
+
+  expect_deprecated(dt_mutate_across(df, where(is.numeric), function(.x) .x + 1))
+  expect_equal(df$x_start, c(2,2,2))
+  expect_equal(df$end_x, c(3,3,3))
+})
+
 test_that("mutate_if() works for numeric columns", {
   df <- data.table(x = c(1,1,1), y = c(2,2,2), z = c("a", "a", "b"))
   df <- df %>%
@@ -82,15 +92,6 @@ test_that("mutate_all() works with data.frame", {
   expect_equal(df$y, c("2","2","2"))
 })
 
-test_that("dt_mutate_across(): .cols works with is.numeric", {
-  df <- data.table(x_start = c(1,1,1), end_x = c(2,2,2), z = c("a", "a", "b"))
-  df <- df %>%
-    dt_mutate_across(where(is.numeric), function(.x) .x + 1)
-
-  expect_equal(df$x_start, c(2,2,2))
-  expect_equal(df$end_x, c(3,3,3))
-})
-
 test_that("mutate_across.(): .cols works with is.numeric", {
   df <- data.table(x_start = c(1,1,1), end_x = c(2,2,2), z = c("a", "a", "b"))
   df <- df %>%
@@ -171,28 +172,28 @@ test_that("mutate_across.() works with twiddle", {
   expect_equal(anon_df, twiddle_df)
 })
 
-test_that("mutate_across() works with by", {
+test_that("mutate_across() works with .by", {
   test_df <- data.table::data.table(
     x = c(1,2,3),
     y = c(4,5,6),
     z = c("a","a","b"))
 
   results_df <- test_df %>%
-    mutate_across.(c(x, y), ~ mean(.x), by = z)
+    mutate_across.(c(x, y), ~ mean(.x), .by = z)
 
   expect_named(results_df, c("x", "y", "z"))
   expect_equal(results_df$x, c(1.5, 1.5, 3))
   expect_equal(results_df$y, c(4.5, 4.5, 6))
 })
 
-test_that("mutate_across.() works with by enhanced selection", {
+test_that("mutate_across.() works with .by enhanced selection", {
   test_df <- data.table::data.table(
     x = c(1,2,3),
     y = c(4,5,6),
     z = c("a","a","b"))
 
   results_df <- test_df %>%
-    mutate_across.(c(x, y), ~ mean(.x), by = where(is.character))
+    mutate_across.(c(x, y), ~ mean(.x), .by = where(is.character))
 
   expect_named(results_df, c("x", "y", "z"))
   expect_equal(results_df$x, c(1.5, 1.5, 3))