+++ tests/testthat/test-slice.R
@@ -1,22 +1,32 @@
-test_that("dt_ works without by", {
+test_that("dt_slice() works without .by", {
   test_df <- tidytable(x = c(1,2,3,4), y = c(4,5,6,7), z = c("a", "a", "a", "b"))
 
   sliced_df <- test_df %>%
     dt_slice(1:4)
 
+  expect_deprecated(dt_slice(test_df, 1:4))
   expect_equal(sliced_df, head(test_df, 4))
 })
 
-test_that("works without by", {
+test_that("works without .by", {
   test_df <- tidytable(x = c(1,2,3,4), y = c(4,5,6,7), z = c("a", "a", "a", "b"))
 
   sliced_df <- test_df %>%
-    slice.(1:4)
+    slice.(1:3)
 
-  expect_equal(sliced_df, head(test_df, 4))
+  expect_equal(sliced_df, head(test_df, 3))
 })
 
-test_that("works without by with .N", {
+test_that("doesn't return NAs", {
+  test_df <- tidytable(x = c(1,2,3,4), y = c(4,5,6,7), z = c("a", "a", "a", "b"))
+
+  sliced_df <- test_df %>%
+    slice.(1:5)
+
+  expect_equal(sliced_df, test_df)
+})
+
+test_that("works without .by with .N", {
   test_df <- tidytable(x = c(1,2,3,4), y = c(4,5,6,7), z = c("a", "a", "a", "b"))
 
   sliced_df <- test_df %>%
@@ -25,7 +35,7 @@ test_that("works without by with .N", {
   expect_equal(sliced_df, head(test_df, 4))
 })
 
-test_that("works without by with n.()", {
+test_that("works without .by with n.()", {
   test_df <- tidytable(x = c(1,2,3,4), y = c(4,5,6,7), z = c("a", "a", "a", "b"))
 
   sliced_df <- test_df %>%
@@ -34,7 +44,7 @@ test_that("works without by with n.()", {
   expect_equal(sliced_df, head(test_df, 4))
 })
 
-test_that("works without by with n.()", {
+test_that("works without .by with n.()", {
   test_df <- tidytable(x = c(1,2,3,4), y = c(4,5,6,7), z = c("a", "a", "a", "b"))
 
   sliced_df <- test_df %>%
@@ -43,7 +53,7 @@ test_that("works without by with n.()", {
   expect_equal(sliced_df, head(test_df, 4))
 })
 
-test_that("works without by with data.frame", {
+test_that("works without .by with data.frame", {
   test_df <- data.frame(x = c(1,2,3,4), y = c(4,5,6,7), z = c("a", "a", "a", "b"))
   sliced_df <- test_df %>%
     slice.(1:4)
@@ -51,41 +61,41 @@ test_that("works without by with data.frame", {
   expect_equal(sliced_df, head(as_tidytable(test_df), 4))
 })
 
-test_that("works with by", {
+test_that("works with .by", {
   test_df <- tidytable(x = c(1,2,3,4), y = c(4,5,6,7), z = c("a", "a", "a", "b"))
   sliced_df <- test_df %>%
-    slice.(1, by = z)
+    slice.(1, .by = z)
 
   expect_equal(sliced_df$z, c("a", "b"))
   expect_equal(sliced_df$x, c(1, 4))
   expect_equal(sliced_df$y, c(4, 7))
 })
 
-test_that("works with by enhanced selection", {
+test_that("works with .by enhanced selection", {
   test_df <- tidytable(x = c(1,2,3,4), y = c(4,5,6,7), z = c("a", "a", "a", "b"))
   sliced_df <- test_df %>%
-    slice.(1, by = where(is.character))
+    slice.(1, .by = where(is.character))
 
   expect_equal(sliced_df$z, c("a", "b"))
   expect_equal(sliced_df$x, c(1, 4))
   expect_equal(sliced_df$y, c(4, 7))
 })
 
-test_that("works with by w/ data.frame", {
+test_that("works with .by w/ data.frame", {
   test_df <- data.frame(x = c(1,2,3,4), y = c(4,5,6,7), z = c("a", "a", "a", "b"),
                         stringsAsFactors = FALSE)
   sliced_df <- test_df %>%
-    slice.(1, by = z)
+    slice.(1, .by = z)
 
   expect_equal(sliced_df$z, c("a", "b"))
   expect_equal(sliced_df$x, c(1, 4))
   expect_equal(sliced_df$y, c(4, 7))
 })
 
-test_that("works with by with data.frame", {
+test_that("works with .by with data.frame", {
   test_df <- tidytable(x = c(1,2,3,4), y = c(4,5,6,7), z = c("a", "a", "a", "b"))
   sliced_df <- test_df %>%
-    slice.(1, by = z)
+    slice.(1, .by = z)
 
   expect_equal(sliced_df$z, c("a", "b"))
   expect_equal(sliced_df$x, c(1, 4))
@@ -97,7 +107,7 @@ test_that("works in custom function", {
 
   slice_fn <- function(.df, val) {
     .df %>%
-      slice.(val, by = z)
+      slice.({{ val }}, .by = z)
   }
 
   sliced_df <- test_df %>%
@@ -126,6 +136,14 @@ test_that("_head.() works when empty", {
   expect_equal(sliced_df, head(test_df, 5))
 })
 
+test_that("_head.() works when empty, doesn't return NAs", {
+  test_df <- tidytable(x = 1:3, y = 1:3)
+  sliced_df <- test_df %>%
+    slice_head.()
+
+  expect_equal(sliced_df, head(test_df, 3))
+})
+
 
 test_that("_head.() works with n specified", {
   test_df <- tidytable(x = 1:10, y = 20:11, z = c(rep("a", 6), rep("b", 4)))
@@ -140,7 +158,7 @@ test_that("_head.() works with n specified", {
 
   datatable_df <- test_df[, head(.SD, 3), by = z]
   sliced_df <- test_df %>%
-    dt_slice_head(n = 3, by = z)
+    dt_slice_head(n = 3, .by = z)
 
   expect_equal(datatable_df, sliced_df)
 })
@@ -180,12 +198,12 @@ test_that("_tail.() works with n specified", {
   expect_equal(sliced_df, test_df)
 })
 
-test_that("_tail() works with n specified with by", {
+test_that("_tail() works with n specified with .by", {
   test_df <- tidytable(x = 1:10, y = 20:11, z = c(rep("a", 6), rep("b", 4)))
 
   datatable_df <- test_df[, tail(.SD, 3), by = z]
   sliced_df <- test_df %>%
-    slice_tail.(n = 3, by = z)
+    slice_tail.(n = 3, .by = z)
 
   expect_equal(datatable_df, sliced_df)
 })
@@ -210,19 +228,19 @@ test_that("_min.() works", {
   expect_equal(sliced_df$y, c(11,12,13))
 })
 
-test_that("_min.() works with by", {
+test_that("_min.() works with .by", {
   test_df <- tidytable(x = 1:10, y = 20:11, z = c(rep("a", 6), rep("b", 4)))
   sliced_df <- test_df %>%
-    slice_min.(order_by = x, n = 3, by = z)
+    slice_min.(order_by = x, n = 3, .by = z)
 
   expect_equal(sliced_df$z, c("a", "a", "a", "b", "b", "b"))
   expect_equal(sliced_df$y, c(20,19,18,14,13,12))
 })
 
-test_that("_min.() works with by enhanced selection", {
+test_that("_min.() works with .by enhanced selection", {
   test_df <- tidytable(x = 1:10, y = 20:11, z = c(rep("a", 6), rep("b", 4)))
   sliced_df <- test_df %>%
-    slice_min.(order_by = x, n = 3, by = where(is.character))
+    slice_min.(order_by = x, n = 3, .by = where(is.character))
 
   expect_equal(sliced_df$z, c("a", "a", "a", "b", "b", "b"))
   expect_equal(sliced_df$y, c(20,19,18,14,13,12))
@@ -233,7 +251,7 @@ test_that("_max.() works with custom function with quosures", {
 
   slice_fn <- function(.df, col, num) {
     .df %>%
-      slice_max.({{ col }}, num)
+      slice_max.({{ col }}, {{ num }})
   }
 
   sliced_df <- test_df %>%