+++ NEWS.md
@@ -1,3 +1,15 @@
+# tidytable 0.5.2
+
+* This version contains general performance improvements
+* Deprecations:
+  + Group by arg `by` has been deprecated for `.by`.
+  This will allow "by" to be used as a column name in `mutate.()` & `summarize.()` in future releases.
+  + All `dt_verb()` functions are soft deprecated
+* New functions:
+  + `separate_rows.()`
+* Functionality improvements:
+  + `unnest.()`: Added `.keep_all` arg to keep all list columns that were not unnested
+
 # tidytable 0.5.1
 
 * New functions:
@@ -57,7 +69,7 @@
   + All functions are now written as `verb.()`
   + `dt_verb()` functions still work, but may be slowly deprecated in future releases
 * Functionality improvements:
-  + Enhanced selection now works in "by"" calls
+  + Enhanced selection now works in "by" calls
   + `filter.()` now works with "by"
   + `pivot_wider.()` can now be used with an aggregation function
   + `pull.()` defaults to last column in a data.table